+++ villagesql/test-extensions/vsql-test-only/src/extension.cc
@@ -595,7 +595,7 @@ void int_sum_all_prerun(vsql::PrerunArgs args, vsql::PrerunResult result) {
 }
 
 void int_sum_all_impl(vsql::VarArgs args, vsql::IntResult out) {
-  long long total = 0;
+  int64_t total = 0;
   for (auto a : args) {
     if (!a.is_null() && a.is_int()) total += a.as_int();
   }
